refactor(a11y): move GoalSemanticDescription component to goal chart specific code area
