+++ packages/charts/src/chart_types/goal_chart/components/goal_semantic_description.tsx
@@ -8,8 +8,8 @@
 
 import React, { Fragment } from 'react';
 
-import type { BandViewModel } from '../../chart_types/goal_chart/layout/types/viewmodel_types';
-import type { A11ySettings } from '../../state/selectors/get_accessibility_config';
+import type { A11ySettings } from '../../../state/selectors/get_accessibility_config';
+import type { BandViewModel } from '../layout/types/viewmodel_types';
 
 interface GoalSemanticDescriptionProps {
   bandLabels: BandViewModel[];
