+++ packages/charts/src/components/accessibility/index.ts
@@ -8,6 +8,5 @@
 
 /* @internal */
 export { ScreenReaderSummary } from './screen_reader_summary';
-export { GoalSemanticDescription } from './goal_semantic_description';
 export { AxisDescriptions } from './axis_descriptions';
 export type { ChartLabelData } from './types';
